Add files via upload
First of all import all the required libraries like pandas, numpy, seaborn, matplotlib etc. And load the data (SALARYPREDICTION.csv file) to train the model.Data stores two features: Experience in Years and Salary. Take the year of experience as an input feature and the salary feature as an output.
We have used -
1)Linear Regression
2) Random Forest